SB-1109: configuration resources storage fixed

+++ strongbox-storage/strongbox-storage-api/src/main/java/org/carlspring/strongbox/booters/ResourcesBooter.java
@@ -73,7 +73,12 @@ public class ResourcesBooter
         for (Resource resource : resources)
         {
             File destFile = new File(configDir, resource.getFilename());
-
+            
+            if (destFile.exists()) {
+                logger.info(String.format("Resource already exists, skip [%s].", destFile));
+                continue;
+            }
+            
             FileUtils.copyInputStreamToFile(resource.getInputStream(), destFile);
         }
     }
